dts: Support app independent but board specific DTS overlays
For any driver development that happens in the internal tree for
boards that are already upstream we need a way to extend the DTS using
an overlay. We do this by taking advantage of the DTC_OVERLAY_FILE
mechanism, while also preserving support for app-specific overlay
files in the app root directory.

Signed-off-by: Johan Hedberg <johan.hedberg@intel.com>