driver: ethernet: dwc_eqos: Fix MetaIRQ race condition on TX
When MetaIRQ is enabled for bottom half interrupt service, the race
condition will happen between the send routine and the send completion
interrupt routine. It is due to MetaIRQ has the highest priority over
the other Zephyr threads ---- send routine. This fix patch ensured the
send routine complete all the necessary sequence before it trigger the
HW DMA to start the send action to prevent the send complete interrupt
intercepting in between the send sequence.

Signed-off-by: Kweh Hock Leong <hock.leong.kweh@intel.com>